+++ pyproject.toml
@@ -123,15 +123,20 @@ exclude = [
 fix = true
 
 [tool.ruff.lint]
+ignore = [
+  "F811",
+  "PLR0911",  # Too many return statements
+  "PLR0912",  # Too many branches
+  "PLR0913",
+]
 select = [
   "E",   # pycodestyle
   "F",   # pyflakes
   "D",   # pydocstyle
   "YTT", # flake8-2020
   "RUF", # Ruff-specific rules
-  "I001",  # isort
+  "I001", # isort
 ]
-fixable = ["I001"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "numpy"
